[BACKLOG-4445] - adding a utility method to copy (convert) one metastore instance to another to support backup/restore of the metastore as part of a full system backup/restore.
I had to modify the XMLMetaStore slightly to support converting between a repository-based metastore and an xml one. it seems a while back there was a refactor for performance reasons to use element type names rather than id's on the filesystem. However, this wasn't fully completed. Since Id's are not null coming from a repository metastore, the resulting xml metastore file system structure was using the UUID's for the element types. This caused a failure when trying to create a new element for a given element type, it couldn't find the folder it was supposed to put the xml file in since it expected the element type name to be the folder, not it's id.